+++ packages/google-vertex/src/google-vertex-options.ts
@@ -23,10 +23,11 @@ export type GoogleVertexModelId =
   | 'gemini-2.0-flash-lite-preview-02-05'
   | 'gemini-2.5-flash-lite-preview-09-2025'
   | 'gemini-2.5-flash-preview-09-2025'
-  | 'gemini-3.1-pro-preview'
   | 'gemini-3-pro-preview'
   | 'gemini-3-pro-image-preview'
   | 'gemini-3-flash-preview'
+  | 'gemini-3.1-pro-preview'
+  | 'gemini-3.1-flash-image-preview'
   // Experimental models
   | 'gemini-2.0-pro-exp-02-05'
   | 'gemini-2.0-flash-exp'