+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @ai-sdk/vue
 
+## 3.0.103
+
+### Patch Changes
+
+- ai@6.0.103
+
+## 3.0.102
+
+### Patch Changes
+
+- ai@6.0.102
+
+## 3.0.101
+
+### Patch Changes
+
+- Updated dependencies [5230482]
+  - ai@6.0.101
+
+## 3.0.100
+
+### Patch Changes
+
+- Updated dependencies [b7fba77]
+  - ai@6.0.100
+
+## 3.0.99
+
+### Patch Changes
+
+- ai@6.0.99
+
+## 3.0.98
+
+### Patch Changes
+
+- ai@6.0.98
+
 ## 3.0.97
 
 ### Patch Changes